Algunos cambios: ================================================================================ CAMBIOS REALIZADOS ================================================================================ Se deja una sola linea glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT); en Game.render
Se simplifica PhysicsManager.renderme():
Se eliminan líneas comentadas de dibujo de sombras y movimiento de cámara dentro de PhysicsManager porque no lo necesitamos.

cStringUtils:.Tokenize para centralizar el código de tokenizado en un método estático.

CAMBIOS: Reordenar data/resources.xml
Se agrupan las líneas de objects player, vehicle, pista y otros.

---------------------------------------------------------------------------------
En resultados aparece lo siguiente:
[cHudManager] XML Load: FAILED
XML Load: FAILED
XML Load: FAILED
[cObjectManager] XML Load: FAILED
XML Load: FAILED
[cRaceControlManager] XML Load: FAILED

Estos mensajes de error se deben a un error en la sintáxis del resources.xml se corrigen lo errores de los tag del Menú. Según parece estos tags no se están
leyendo todavía por lo que se podrían comentar.

---------------------------------------------------------------------------------
Render de OpenGL: glEnable(GL_CULL_FACE); no estaba activo hay que activarlo al renderizar los modelos y desactivarlo para que se vea el HUD¿¿??

Corrigiendo warnings de SoundManager, ObjectManager, etc.

se deshabilita Lua, Characters-lua, Behaviour-lua del proyecto porque no los vamos a usar.

SE CAMBIA: Las variables enum de Scenes se renombran según notación Húngara Se añade eGameplay a Gameplay. Esto facilita la lectura del código.

Se CAMBIA un poco el cálculo de FPS, una vez que se calcula el FPS se pone la variable tiempo acumulado a cero para empezar a calcular sin arrastrar valores del periodo 
anterior. Cambiado en FPSCounter.Update().

se crea clase cTimeDebug, para intentar medir los tiempos de zonas de código... hay que depurarlo porque el imprimir los mensajes de tiempo en consola retardan la
ejecución y se muestran valores bajos a gran velocidad. Lo cual no es muy útil.
Tengo la idea de mostrar valores acumulados cada x segundos, pero en usando OutputDebugString a ver si el resultado es más práctico.

¿les parece buena idea tener en el juego varias posiciones de cámara? Total... ya lo tenemos hecho.

Quitar ControlRaceManager del proyecto y dejar RaceControlManager.

---------------------------------------------------------------------------------
PROBLEMA: No se puede abrir "InstaladorJuego.vdproj". Tipo de proyecto no compatible con esta versión de la aplicación.
SUGERENCIA: Eliminar del proyecto y del svn, o subir los ficheros...

PROBLEMA (sólo de David): Línea 249 de mesh.cpp
Mesh.cpp [./Data/Circuitos/Basico/BasicoMuros2.DAE_0] ERROR OpenGL: 1282
De momento he puesto una línea #ifdef _DAVID para controlar este error que sólo tengo yo.

+++ trunk/BasicEngine/BasicEngine/Utility/ResourceManager.cpp
@@ -104,7 +104,7 @@ cResourceHandle cResourceManager::FindResourceIndice( int liIndice )
 	cResourceHandle lHandle;
 	
 	//for ( unsigned luiIndex = 0; luiIndex < muiMaxSize; ++luiIndex )
-	if (liIndice < muiMaxSize)
+	if (liIndice < (int) muiMaxSize)
 	{
 			// Is a valid resource?
 		if ( maResources[liIndice].muiKey != kuiInvalidKey )
